requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
albucore==0.0.24
albumentations==2.0.8
alphashape==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.5
bio==1.8.0
biopython==1.85
biothings_client==0.4.1
boto3==1.40.30
botocore==1.40.30
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
chess==1.11.2
click==8.2.1
click-log==0.4.0
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
# Crawl4AI==0.7.4
cryptography==45.0.7
curl_cffi==0.13.0
cycler==0.12.1
# ddgs==9.5.5
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
exceptiongroup==1.3.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fast-langdetect==0.2.5
fastapi==0.116.1
fasttext-predict==0.9.2.4
fastuuid==0.12.0
ffmpy==0.6.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
google==3.0.0
google-auth==2.40.3
gprofiler-official==1.0.0
gradio==5.45.0
gradio_client==1.13.0
gradio_pdf==0.0.22
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json_repair==0.50.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
lark==1.2.2
lazy_loader==0.4
litellm==1.77.0
loguru==0.7.3
lxml==5.4.0
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mineru==2.2.2
mmh3==5.2.0
modelscope==1.29.2
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.12
mygene==3.2.2
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
# nvidia-cublas-cu12==12.6.4.1
# nvidia-cuda-cupti-cu12==12.6.80
# nvidia-cuda-nvrtc-cu12==12.6.77
# nvidia-cuda-runtime-cu12==12.6.77
# nvidia-cudnn-cu12==9.5.1.17
# nvidia-cufft-cu12==11.3.0.4
# nvidia-cufile-cu12==1.11.1.6
# nvidia-curand-cu12==10.3.7.77
# nvidia-cusolver-cu12==11.7.1.2
# nvidia-cusparse-cu12==12.5.4.2
# nvidia-cusparselt-cu12==0.6.3
# nvidia-nccl-cu12==2.26.2
# nvidia-nvjitlink-cu12==12.6.85
# nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.107.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
patchright==1.55.1
pdfminer.six==20250506
pdftext==0.6.3
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
polars==1.33.1
pooch==1.8.2
primp==0.15.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
PubChemPy==1.0.5
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.4
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
robust-downloader==0.0.2
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
ruff==0.13.0
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
selenium==4.35.0
semantic-version==2.10.0
serpapi==0.1.5
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
-e git+https://github.com/huggingface/smolagents.git@e106fce649b69b6c7857c89443921f75f9e36378#egg=smolagents
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
stringzilla==4.0.10
sympy==1.14.0
tenacity==9.1.2
tf-playwright-stealth==1.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.11.0
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.56.1
trimesh==4.8.1
trio==0.30.0
trio-websocket==0.12.2
# triton==3.3.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.200
ultralytics-thop==2.0.17
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
zstandard==0.25.0
